Merge develop on feature v3

+++ tools/clean.sh
@@ -5,6 +5,13 @@
 
 set -e
 source ./tools/utils/echo-color.sh
+source ./tools/utils/argparse.sh
+
+set_description "Cleans up the local environment."
+define_arg "derived-data" "false" "Clean Xcode derived data." "store_true"
+define_arg "pods" "false" "Clean Pods." "store_true"
+define_arg "xcconfigs" "false" "Clean local xcconfigs." "store_true"
+define_arg "carthage" "false" "Clean Carthage cache." "store_true"
 
 clean_dir() {
     local dir="$1"
@@ -16,13 +23,29 @@ clean_dir() {
     fi
 }
 
-clean_dir ~/Library/Developer/Xcode/DerivedData
-clean_dir ~/Library/Caches/org.carthage.CarthageKit/dependencies/
-clean_dir ~/Library/org.swift.swiftpm
-clean_dir ./Carthage/Build
-clean_dir ./Carthage/Checkouts
-clean_dir ./IntegrationTests/Pods
+check_for_help "$@"
+parse_args "$@"
 
-echo_warn "Cleaning local xcconfigs"
-rm -vf ./xcconfigs/Base.ci.local.xcconfig
-rm -vf ./xcconfigs/Base.dev.local.xcconfig
+if [[ "$derived-data" == "true" ]]; then
+    echo_subtitle "Cleaning Xcode derived data"
+    clean_dir ~/Library/Developer/Xcode/DerivedData
+    clean_dir ~/Library/org.swift.swiftpm
+fi
+
+if [[ "$pods" == "true" ]]; then
+    echo_subtitle "Cleaning Pods"
+    clean_dir ./IntegrationTests/Pods
+fi
+
+if [[ "$xcconfigs" == "true" ]]; then
+    echo_subtitle "Cleaning local xcconfigs"
+    rm -vf ./xcconfigs/Base.ci.local.xcconfig
+    rm -vf ./xcconfigs/Base.dev.local.xcconfig
+fi
+
+if [[ "$carthage" == "true" ]]; then
+    echo_subtitle "Cleaning Carthage cache"
+    clean_dir ~/Library/Caches/org.carthage.CarthageKit/dependencies/
+    clean_dir ./Carthage/Build
+    clean_dir ./Carthage/Checkouts
+fi
